[MODIFY] </main/../payload/requests/EquipmentApplyRequest.java>
mv request/ requests/
